back/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
motor==3.3.1
redis==5.0.1
aiogram>=3.4.0   # FRONTEND BOT
python-telegram-bot # WORKERS/TASKS SENDING
pillow==10.1.0
google-generativeai==0.3.1
yookassa==2.3.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.1
